requirements.txt
analytics-python==1.2.9
Faker==1.0.2

